refactor: getStations 로직 Sections로 이동
